backend/requirements.txt
Flask>=2.2.5
Flask-Cors>=4.0,<5.0
Flask-SocketIO>=5.3,<6.0
python-socketio>=5.8,<5.10

